hatico_service/backend/api/requirements.txt
anyio==3.4.0
asgiref==3.4.1
asynclog==0.1.7
bson==0.5.10
click==8.0.3
fastapi==0.70.0
h11==0.12.0
idna==3.3
importlib-metadata==4.8.2
loguru==0.5.3
pydantic==1.8.2
pymongo==3.12.1
python-dateutil==2.8.2
python-dotenv==0.19.2
python-multipart==0.0.5
six==1.16.0
sniffio==1.2.0
starlette==0.16.0
typing-extensions==4.0.0
uvicorn==0.15.0
zipp==3.6.0
